airflow/requirements.txt
requests
requests-html
selectolax
xlrd
google-cloud-bigquery
google-cloud-storage
pandas
yfinance
numpy
scipy

